+++ assets/levelup/LevelUpHUD.java
@@ -5,6 +5,7 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.Gui;
 import net.minecraft.client.gui.ScaledResolution;
 import net.minecraft.util.StatCollector;
+import net.minecraftforge.client.event.FOVUpdateEvent;
 import net.minecraftforge.client.event.RenderGameOverlayEvent;
 import net.minecraftforge.client.event.RenderGameOverlayEvent.ElementType;
 
@@ -13,9 +14,11 @@ import java.util.List;
 
 public final class LevelUpHUD extends Gui {
     public static final LevelUpHUD INSTANCE = new LevelUpHUD();
-    private float val = 0.7F, valIncr = 0.005F;
+    private float val, valIncr;
 
     private LevelUpHUD() {
+        val = 0.7F;
+        valIncr = 0.005F;
     }
 
     public void addToText(List<String> left) {
@@ -36,11 +39,27 @@ public final class LevelUpHUD extends Gui {
 
     @SubscribeEvent
     public void renderLvlUpHUD(RenderGameOverlayEvent.Pre event) {
-        if (LevelUp.proxy.getPlayer() != null) {
+        if (LevelUp.allowHUD && LevelUp.proxy.getPlayer() != null) {
             if (LevelUp.renderTopLeft && event.type == ElementType.TEXT)
                 addToText(((RenderGameOverlayEvent.Text) event).left);
-            if (LevelUp.renderExpBar && event.type == ElementType.EXPERIENCE) {
+            if (LevelUp.renderExpBar && event.type == ElementType.EXPERIENCE)
                 addToExpBar(event.resolution);
+        }
+    }
+
+    @SubscribeEvent
+    public void onFOV(FOVUpdateEvent event){
+        if(!LevelUp.changeFOV && !event.entity.isUsingItem()) {
+            int skill = 0;
+            if(event.entity.isSneaking()){
+                skill = 2 * FMLEventHandler.getSkill(event.entity, 8);
+            }else if(event.entity.isSprinting()){
+                skill = FMLEventHandler.getSkill(event.entity, 6);
+            }
+            if(skill > 0){
+                event.newfov -= 0.5F;
+                event.newfov *=  1/(1.0F + skill / 100F);
+                event.newfov += 0.5F;
             }
         }
     }